do try not remove /bin/sh when $TMPDIR not yet set; prevents confusing error message (and trouble when ran as root)
git-svn-id: https://secure.a-eskwadraat.nl/svn/domjudge/trunk@1678 caf94f87-7bd6-0310-88dc-dee8202e2269

+++ judge/test_solution.sh
@@ -64,7 +64,9 @@ cleanexit ()
 	fi
 
 	# Remove copied static shell to save disk space
-	rm -f "$TMPDIR/bin/sh"
+	if [ "$TMPDIR" ]; then
+		rm -f "$TMPDIR/bin/sh"
+	fi
 
 	logmsg $LOG_INFO "exiting"
 }
